pyproject.toml
[tool.poetry]
name = "unitree"
version = "0.1.0"
description = ""
authors = ["snorkysnark <snorkysnark@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.3.1"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
jinja2 = "^3.1.4"
returns = "^0.22.0"


[tool.poetry.group.dev.dependencies]
jupyter-console = "^6.6.3"
debugpy = "^1.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
